Create C++  Space-managment
A C++ program that simulates simple management of virtual space resources. The program allows you to reserve, release and check the availability of resources in the virtual space, move resources, check the status of a specific resource, exchange reserved and available resources, as well as generate unique identifiers for resources.